Fix #52408 and #75075 
Will add related tests after deciding correct format, thanks !