@lateapexearlyspeed 2022-09-07
Hi @krwq just confirm, I will change code to implement just [proposal description](https://github.com/dotnet/runtime/issues/52408#issue-877928403) and align one [overload behavior issue](https://github.com/dotnet/runtime/issues/75075#issue-1361510964), and keep all other existing behavior no change.

I have pushed one version of prod code, and will continue adding tests later. If possible, maybe you can review and point out issue in advance, thanks !