@lateapexearlyspeed 2022-09-15
Hi @krwq Tests added, please review, thanks!

They used newtonsoft.json to do assertion as existing cases, but have to add some compensation newlines and whitespaces so that can verify new cases. Note the compensations are not entirely for new proposal, partly it is because we lost some test cases for existing indented behaviors of comment previously.