@krwq 2022-11-04
Any chance you could run perf tests on this change? There are changes made to the hot path so it would be useful to know if we regressed something or not and if yes how can we mitigate this