@lateapexearlyspeed 2022-11-09
> Let me know if you're planning to fix this or not

Changed, please check.

> Any chance you could run perf tests

I would try to run perf test after deciding current json format is ok, please help confirm, thanks @krwq !